backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-session==0.5.0
google-generativeai==0.3.2
werkzeug==3.0.1
